Int64 -> Int

+++ base/loading.jl
@@ -910,9 +910,9 @@ function modulnamehint(tn::AbstractString)
     hns = vcat(try collect(keys(get(parsed_toml(load_path()[1]), "deps", nothing)))
             catch; String[] end, isdir(Sys.STDLIB) ? readdir(Sys.STDLIB) : String[])
     ((length(tn) < 3) || isempty(hns)) && return nothing
-    scores = similar(hns, Int64)
+    scores = similar(hns, Int)
     for (index, hn) in enumerate(hns)
-        (length(tn) >= length(hn)) ? (o1 = tn ; o2 = hn) : (o1 = hn; o2 = tn)
+        (length(tn) >= length(hn)) ? (o1 = tn; o2 = hn) : (o1 = hn; o2 = tn)
         scores[index] = length(o1) - length(o2) + (contains(o1, o2) ? 0 :
                         contains(lowercase(o1), lowercase(o2)) ? 1 : 4)
     end
